Allow a restart file to be loaded directly in main
not necessarily passing through ARC.py, so API could use input/restart
files
